+++ src/main/java/com/yunhuakeji/attendance/constants/ConfigConstants.java
@@ -2,8 +2,8 @@ package com.yunhuakeji.attendance.constants;
 
 public class ConfigConstants {
 
-    public static final Short CHECK_DEVICE_YES = 1;
-    public static final Short CHECK_DEVICE_NO = 2;
+    public static final Byte CHECK_DEVICE_YES = 1;
+    public static final Byte CHECK_DEVICE_NO = 2;
 
     //1未打卡，2已打卡
     public static final byte INSTRUCTOR_NOT_CLOCK = 1;
